requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
langchain-community==0.0.10
openai==1.3.5
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
pypdf==3.17.1
nbformat==5.9.2
python-dotenv==1.0.0
aiohttp==3.9.1 